Установочные скрипты

+++ install.py
@@ -2,6 +2,7 @@
 
 import os,shutil
 import sys
+import argparse
 
 def make_path(path):
   if not os.path.exists(path):
@@ -9,11 +10,11 @@ def make_path(path):
 
 def get_files(prefix):
   files={
-    'mcgdb-mcedit'     :('mc/src/mcgdb-mc'  ,   '{}/bin/mcgdb-mcedit'.format(prefix)                ,   0777),
-    'mcgdb'            :('mcgdb'            ,   '{}/bin/mcgdb'.format(prefix)                       ,   0777),
+    'mcgdb-mcedit'     :('obj-mc/src/mc'    ,   '{}/bin/mcgdb-mcedit'.format(prefix)                ,   0777),
+    'mcgdb'            :(None               ,   '{}/bin/mcgdb'.format(prefix)                       ,   0777),
     'mcgdb.py'         :('mcgdb.py'         ,   '{}/share/mcgdb/mcgdb.py'.format(prefix)            ,   0777),
     'defines-mcgdb.gdb':('defines-mcgdb.gdb',   '{}/share/mcgdb/defines-mcgdb.gdb'.format(prefix)   ,   0777),
-    'startup.gdb'      :('startup.gdb'      ,   '{}/share/mcgdb/startup.gdb'.format(prefix)         ,   0777),
+    'startup.gdb'      :(None               ,   '{}/share/mcgdb/startup.gdb'.format(prefix)         ,   0777),
   }
   return files
 
@@ -24,27 +25,40 @@ def install(prefix):
   files=get_files(prefix)
   for fname in files:
     src,dst,mode=files[fname]
+    if src==None:
+      continue
     if os.path.exists(dst):
       os.remove(dst)
     shutil.copy(src,dst)
     os.chmod(dst,mode)
-  os.system("sed 's/PATH_TO_MC=.*/PATH_TO_MC={mcedit}/' {dst}".format(
+  os.system('''sed 's#PATH_TO_MC=.*#PATH_TO_MC="{mcedit}"#' {dst} -i'''.format(
     dst=files['mcgdb.py'][1],
     mcedit=files['mcgdb-mcedit'][1],
   ))
-  os.system("sed 's/PATH_TO_DEFINES_MCGDB=.*/PATH_TO_DEFINES_MCGDB={defines}/' {dst}".format(
+  os.system('''sed 's#PATH_TO_DEFINES_MCGDB=.*#PATH_TO_DEFINES_MCGDB="{defines}"#' {dst} -i'''.format(
     dst=files['mcgdb.py'][1],
     defines=files['defines-mcgdb.gdb'][1],
   ))
+  with open(files['startup.gdb'][1],'w') as f:
+    f.write('''
+source {mcgdb_py}
+pi mc()
+pi mcgdb_main_window()
+'''.format(mcgdb_py=files['mcgdb.py'][1]))
+  with open(files['mcgdb'][1],'w') as f:
+    f.write('''#!/usr/bin/env bash
+gdb $@ -x {startup}
+'''.format(startup=files['startup.gdb'][1]))
 
 
 def remove(prefix):
   pass
 
 def main():
-  prefix=''
-  if len(sys.argv)>=3 and sys.argv[1]=='--prefix':
-    prefix=sys.argv[2]
+  parser = argparse.ArgumentParser()
+  parser.add_argument("--prefix")
+  args=parser.parse_args()
+  prefix=args.prefix
   install(prefix)
 
 if __name__ == "__main__":
